New style - each class in diff class
Wrong understanding Laofre book it seems! Early i thought - all classes might be in one class (file)
Now - all in different in one package.
